make pivots private
