+++ Source/ActivityPub.Types/Conversion/Converters/TypeMapConverter.cs
@@ -6,7 +6,6 @@ using System.Text.Json.Nodes;
 using System.Text.Json.Serialization;
 using ActivityPub.Types.AS;
 using ActivityPub.Types.Conversion.Overrides;
-using ActivityPub.Types.Conversion.Pivots;
 using ActivityPub.Types.Internal;
 using ActivityPub.Types.Util;
 using JetBrains.Annotations;
@@ -15,32 +14,19 @@ using Microsoft.Extensions.Options;
 namespace ActivityPub.Types.Conversion.Converters;
 
 /// <inheritdoc />
-public partial class TypeMapConverter : JsonConverter<TypeMap>
+public class TypeMapConverter : JsonConverter<TypeMap>
 {
     private readonly IASTypeInfoCache _asTypeInfoCache;
     private readonly IConversionOptions _conversionOptions;
-    private readonly INamelessEntityPivot _namelessEntityPivot;
-    private readonly IAnonymousEntityPivot _anonymousEntityPivot;
-    private readonly ICustomConvertedEntityPivot _customConvertedEntityPivot;
+    private readonly NamelessEntityPivot _namelessEntityPivot = new();
+    private readonly AnonymousEntityPivot _anonymousEntityPivot = new();
+    private readonly CustomConvertedEntityPivot _customConvertedEntityPivot = new();
 
     /// <inheritdoc />
     [UsedImplicitly]
     public TypeMapConverter(IASTypeInfoCache asTypeInfoCache, IOptions<ConversionOptions> conversionOptions)
     {
         _asTypeInfoCache = asTypeInfoCache;
-        _anonymousEntityPivot = new AnonymousEntityPivot();
-        _namelessEntityPivot = new NamelessEntityPivot();
-        _customConvertedEntityPivot = new CustomConvertedEntityPivot();
-        _conversionOptions = conversionOptions.Value;
-    }
-
-    /// <inheritdoc />
-    public TypeMapConverter(IASTypeInfoCache asTypeInfoCache, IOptions<ConversionOptions> conversionOptions, IAnonymousEntityPivot anonymousEntityPivot, INamelessEntityPivot namelessEntityPivot, ICustomConvertedEntityPivot customConvertedEntityPivot)
-    {
-        _asTypeInfoCache = asTypeInfoCache;
-        _anonymousEntityPivot = anonymousEntityPivot;
-        _namelessEntityPivot = namelessEntityPivot;
-        _customConvertedEntityPivot = customConvertedEntityPivot;
         _conversionOptions = conversionOptions.Value;
     }
